+++ docs/changelog.rst
@@ -5,6 +5,36 @@ Changelog
 This page contains a summary of changes between the official cwEPR releases. Only the biggest changes are listed here. A complete and detailed log of all changes is available through the `GitHub Repository Browser <https://github.com/tillbiskup/cwepr/commits/master>`_.
 
 
+Version 0.2.0
+=============
+
+Not yet released
+
+New features
+------------
+
+
+Changes
+-------
+
+
+Fixes
+-----
+
+
+Version 0.1.1
+=============
+
+Not yet released
+
+
+The following bugs have been fixed:
+
+* Bugfix in Normalisation in combination with aspecd (afaa3c79)
+* Bugfix in Magnettech-Import, additional test for InfofileReporter (36ea8257)
+* Bugfix in GoniometerSweepImporter and Reporter to get correct format of some numbers (5ab46142)
+
+
 Version 0.1.0
 =============